apps/track_a_b2c_product_advisor/requirements.txt

openai
langchain
faiss-cpu
tiktoken
pandas
python-dotenv
